runners/Label Studio/ML_Backend/requirements.txt
label-studio-ml==1.0.9
torch>=1.13.0
torchvision>=0.14.0
opencv-python>=4.7.0
Pillow>=9.0.0
numpy>=1.21.0
requests>=2.28.0
Flask==1.1.2
ultralytics
redis
rq